Switch from a react app to a node app
