--- src/redux/types.js
@@ -1,5 +1,0 @@
-const START_CRAWLER = 'redux/START_CRAWLER';
-
-export default {
-  START_CRAWLER,
-};
